@@ -10,7 +10,7 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": ">=7.3"
+        "php": ">=8.0"
     },
     "autoload": {
         "files": [
